Fix inconsistent text selection contrast (fixes #31)
Selections in VS Code’s integrated terminal are given a low opacity,
while in the editor they keep their full opacity. This means that using
the same colour as the editor and terminal selection colour results in
varying contrast. This commit purposefully darkens the colour of
selections in the terminal to make them have the same contrast with the
background as selections in the editor do.